Refactor: Two implementations and one service
