+++ news/1 Enhancements/9665.md
@@ -1,0 +1,1 @@
+DataFrame viewer enabled on the web.
