+++ src/platform/common/utils/localize.ts
@@ -1393,12 +1393,8 @@ export namespace DataScience {
         );
     export const listOfFilesWithLinksThatMightNeedToBeRenamed = () =>
         localize('DataScience.listOfFilesWithLinksThatMightNeedToBeRenamed', 'File(s): {0} might need to be renamed.');
-    export const interpreterNotSupportedOnTheWeb = () =>
-        localize('DataScience.interpreterNotSupportedOnTheWeb', 'Interpreter not supported on the web.');
     export const noActiveKernelSession = () =>
         localize('DataScience.noActiveKernelSession', 'No no active kernel session.');
-    export const insufficientParameters = () =>
-        localize('DataScience.insufficientParameters', 'Insufficient parameters.');
     export const failedToGetVersionOfPandas = () =>
         localize(
             { key: 'DataScience.failedToGetVersionOfPandas', comment: ['{Locked="Pandas"}'] },
