+++ src/webviews/extension-side/dataviewer/dataViewerCommandRegistry.ts
@@ -99,10 +99,7 @@ export class DataViewerCommandRegistry implements IExtensionSingleActivationServ
                         this.debugService.activeDebugSession.configuration
                     );
 
-                    pythonEnv &&
-                        (await this.dataViewerDependencyService.checkAndInstallMissingDependencies({
-                            interpreter: pythonEnv
-                        }));
+                    pythonEnv && (await this.dataViewerDependencyService.checkAndInstallMissingDependencies(pythonEnv));
                 }
 
                 const variable = convertDebugProtocolVariableToIJupyterVariable(
