+++ src/webviews/extension-side/dataviewer/types.ts
@@ -7,8 +7,8 @@ import { IKernel } from '../../../kernels/types';
 import { IJupyterVariable } from '../../../kernels/variables/types';
 import { IDisposable } from '../../../platform/common/types';
 import { SharedMessages } from '../../../messageTypes';
-import { PythonEnvironment } from '../../../platform/pythonEnvironments/info';
 import { SliceOperationSource } from '../../../platform/telemetry/constants';
+import { PythonEnvironment } from '../../../platform/pythonEnvironments/info';
 
 export const CellFetchAllLimit = 100000;
 export const CellFetchSizeFirst = 100000;
@@ -129,15 +129,7 @@ export interface IJupyterVariableDataProviderFactory {
 }
 
 export const IDataViewerDependencyService = Symbol('IDataViewerDependencyService');
-export type IDataViewerDependencyServiceOptions =
-    | {
-          interpreter: PythonEnvironment;
-          kernel?: never;
-      }
-    | {
-          interpreter?: never;
-          kernel: IKernel;
-      };
+
 export interface IDataViewerDependencyService {
-    checkAndInstallMissingDependencies(options: IDataViewerDependencyServiceOptions): Promise<void>;
+    checkAndInstallMissingDependencies(executionEnvironment: IKernel | PythonEnvironment): Promise<void>;
 }
